bluetoothHandler: fix segfault with early starting
This commit is trying to fix an early segfault.
If the bluetooth stack is not ready, then the `Powered` value is null.
Calling the `set_value` in this stage is causing `Segmentation fault`.

This was leading to problems with automatic AA start on the Pi,
so it was often necessary to start it manually afterwards:
`/etc/init.d/S93aawgd start` (the process was not running)